Added plugin.go to handle the dynamic loading and usage of the plugin implementation
